+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # Unreleased
 
 * Initial commit
+* Move jitsi download to AMI build